special el handlers impl is now appled after the children are morphed.
