+++ src/components/m-dropdown/js/drop-down.js
@@ -149,17 +149,6 @@ class DropDown extends UiEvents {
       lastElementChild.style.height = '';
       remove(node, this.options.isOpenClass);
     }
-    const select = this.wcNode.querySelector(`.${this.options.nativeSelectClass}`);
-    if (select) {
-      const selectedOption = this.wcNode.querySelector(this.options.selectedItemSelector);
-      if (selectedOption) {
-        if (selectedOption.index !== select.selectedIndex) {
-          select.selectedIndex = selectedOption.index;
-        }
-      } else {
-        select.selectedIndex = 0;
-      }
-    }
   }
 
   destroy() {
