+++ src/js/abstract/utils/morph.js
@@ -7,7 +7,7 @@ export default morph;
 // diff elements and apply the resulting patch to the old node
 // (obj, obj) -> null
 function morph(newNode, oldNode) {
-  const { nodeType, nodeName } = newNode;
+  const { nodeType } = newNode;
 
   if (nodeType === ELEMENT_NODE) {
     copyAttrs(newNode, oldNode);
@@ -15,16 +15,6 @@ function morph(newNode, oldNode) {
 
   if ((nodeType === TEXT_NODE || nodeType === COMMENT_NODE) && oldNode.nodeValue !== newNode.nodeValue) {
     oldNode.nodeValue = newNode.nodeValue;
-  }
-
-  // Some DOM nodes are weird
-  // https://github.com/patrick-steele-idem/morphdom/blob/master/src/specialElHandlers.js
-  if (nodeName === 'INPUT') {
-    updateInput(newNode, oldNode);
-  } else if (nodeName === 'OPTION') {
-    updateOption(newNode, oldNode);
-  } else if (nodeName === 'TEXTAREA') {
-    updateTextarea(newNode, oldNode);
   }
 }
 
@@ -105,69 +95,6 @@ function copyAttrs(newNode, oldNode) {
       } else if (!newNode.hasAttributeNS(null, attrName)) {
         oldNode.removeAttribute(attrName);
       }
-    }
-  }
-}
-
-function updateOption(newNode, oldNode) {
-  updateAttribute(newNode, oldNode, 'selected');
-}
-
-// The "value" attribute is special for the <input> element since it sets the
-// initial value. Changing the "value" attribute without changing the "value"
-// property will have no effect since it is only used to the set the initial
-// value. Similar for the "checked" attribute, and "disabled".
-function updateInput(newNode, oldNode) {
-  const { value: newValue } = newNode;
-  const { value: oldValue } = oldNode;
-
-  updateAttribute(newNode, oldNode, 'checked');
-  updateAttribute(newNode, oldNode, 'disabled');
-
-  if (newValue !== oldValue) {
-    oldNode.setAttribute('value', newValue);
-    oldNode.value = newValue;
-  }
-
-  if (newValue === 'null') {
-    oldNode.value = '';
-    oldNode.removeAttribute('value');
-  }
-
-  if (!newNode.hasAttributeNS(null, 'value')) {
-    oldNode.removeAttribute('value');
-  } else if (oldNode.type === 'range') {
-    // this is so elements like slider move their UI thingy
-    oldNode.value = newValue;
-  }
-}
-
-function updateTextarea(newNode, oldNode) {
-  const { value: newValue } = newNode;
-
-  if (newValue !== oldNode.value) {
-    oldNode.value = newValue;
-  }
-
-  if (oldNode.firstChild && oldNode.firstChild.nodeValue !== newValue) {
-    // Needed for IE. Apparently IE sets the placeholder as the
-    // node value and vise versa. This ignores an empty update.
-    if (newValue === '' && oldNode.firstChild.nodeValue === oldNode.placeholder) {
-      return;
-    }
-
-    oldNode.firstChild.nodeValue = newValue;
-  }
-}
-
-function updateAttribute(newNode, oldNode, name) {
-  if (newNode[name] !== oldNode[name]) {
-    oldNode[name] = newNode[name];
-
-    if (newNode[name]) {
-      oldNode.setAttribute(name, '');
-    } else {
-      oldNode.removeAttribute(name);
     }
   }
 }
